refactor(ai-agent): move mode/model handlers into AgentService
